playlist fragment review fix

+++ app/src/main/java/com/example/playlistmaker/data/favorites/dto/FavoriteEntity.kt
@@ -10,7 +10,7 @@ data class FavoriteEntity(
     @ColumnInfo val itunesId: Int,
     val trackName: String?,
     val artistName: String?,
-    val trackTime: Int,
+    val trackTime: Long,
     val trackTimeFormat: String,
     val artworkUrl100: String?,
     val previewUrl: String,
