+++ app/src/main/java/com/example/playlistmaker/domain/playlists/models/Playlist.kt
@@ -8,6 +8,6 @@ data class Playlist(
     val description: String?,
     var coverUri: String?,
     var trackCount: Int = 0,
-    var duration: Int = 0,
+    var duration: Long = 0,
     var durationFormat: String = "0"
 ) : Serializable
